+++ catalyst-gateway-crates/c509-certificate/src/c509_general_names/other_name_hw_module.rs
@@ -1,0 +1,54 @@
+//! OtherNameHardwareModuleName, special type for hardwareModuleName type of otherName.
+//!  When 'otherName + hardwareModuleName' is used, then [ ~oid, bytes ] is
+//! used to contain the pair ( hwType, hwSerialNum ) directly as specified in [RFC4108] 
+
+use asn1_rs::Oid;
+use minicbor::{encode::Write, Decode, Decoder, Encode, Encoder};
+
+use crate::c509_oid::C509oid;
+
+/// A struct represents the hardwareModuleName type of otherName.
+/// Containing a pair of ( hwType, hwSerialNum ) as mentioned in
+/// [RFC4108](https://datatracker.ietf.org/doc/rfc4108/)
+///
+/// # Fields
+/// * `hw_type` - The hardware type OID.
+/// * `hw_serial_num` - The hardware serial number represent in bytes.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub(crate) struct OtherNameHardwareModuleName {
+    hw_type: C509oid,
+    hw_serial_num: Vec<u8>,
+}
+
+impl OtherNameHardwareModuleName {
+    /// Create a new instance of OtherNameHardwareModuleName.
+    pub fn new(hw_type: Oid<'static>, hw_serial_num: Vec<u8>) -> Self {
+        Self {
+            hw_type: C509oid::new(hw_type),
+            hw_serial_num,
+        }
+    }
+}
+
+impl Encode<()> for OtherNameHardwareModuleName {
+    fn encode<W: Write>(
+        &self, e: &mut Encoder<W>, ctx: &mut (),
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        e.array(2)?;
+        self.hw_type.encode(e, ctx)?;
+        e.bytes(&self.hw_serial_num)?;
+        Ok(())
+    }
+}
+
+impl<'a> Decode<'a, ()> for OtherNameHardwareModuleName {
+    fn decode(d: &mut Decoder<'a>, ctx: &mut ()) -> Result<Self, minicbor::decode::Error> {
+        d.array()?;
+        let hw_type = C509oid::decode(d, ctx)?;
+        let hw_serial_num = d.bytes()?.to_vec();
+        Ok(OtherNameHardwareModuleName::new(
+            hw_type.get_oid(),
+            hw_serial_num,
+        ))
+    }
+}
